// src/generated/models/CreateJobDefinitionEndpoint_Output.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * The result when a job definition has been created.
 */
export type CreateJobDefinitionEndpoint_Output = {
    /**
     * The id of the created job definition.
     */
    definitionId: string;
    /**
     * The id of the created preview pipeline.
     */
    pipelineId: string;
};