// src/generated/models/CreateJobEndpoint_Output.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * The result when a job has been created.
 */
export type CreateJobEndpoint_Output = {
    /**
     * The id of the created job.
     */
    jobId: string;
    /**
     * Whether recruiting was automatically started for the audience.
     */
    recruitingStarted: boolean;
};