// src/generated/models/CreateJobRevisionEndpoint_Output.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * The result when a job revision has been created.
 */
export type CreateJobRevisionEndpoint_Output = {
    /**
     * The revision number of the created job revision.
     */
    revisionNumber: number;
    /**
     * The id of the created preview pipeline.
     */
    pipelineId: string;
};